Implement ZFS System Attributes
This now allows us to read the length information for ZPL files, so we
don't have null bytes at the end anymore!

The way the system attribute registry is implemented currently is a
little funky, but it works. The error handling could be improved and
hopefully the API could be updated so every time we read a file's
attributes we don't copy all the attribute name strings.

zfs-cli can now be used to list a directory's childrens or the contents
of a file, depending on what it finds at the passed path